WIP: Witness serde stuff
